+++ ui/CHANGES.md
@@ -4,6 +4,7 @@
 - **Added**: `VisTextField#setCurosrAtTextEnd()`, `#getProgrammaticChangeEvents()`
 - **Added**: `FileChooser` file sorting options available under right click menu
 - **Added**: `FileChooser#setSorting(FileSorting)` and `#setSortingOrderAscending` along with appropriate getters
+- **Added**: `FileChooser#setSaveLastDirectory` - allows to automatically remember last directory user browsed between app launches, disabled by default
 - **Fixed**: VisSplitPane was not restoring default cursor when user dragged pointer outside od pane area 
 - **Fixed**: [#188](https://github.com/kotcrab/vis-editor/issues/188) - same instance of `VisDialog` couldn't be closed for the second time using close button 
 - **Fixed**: `FileChooser` NPE when user right clicked last file item after deleting all others files